requirements.txt
streamlit>=1.28.0
requests>=2.31.0
spacy>=3.7.0
fuzzywuzzy[speedup]>=0.18.0
joblib>=1.3.0
tinydb>=4.8.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
python-levenshtein>=0.21.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
